+++ changelog.md
@@ -2,6 +2,17 @@
 
 ## [Unreleased]
 
+### Added
+
+-   [Input] Add new property `fullWidth`.
+-   [Input] Add new property `inputRef`.
+
+### Changed
+
+-   [Input] Rename `hasError` property to `invalid`.
+-   [Input] Remove `event` argument from `onValueChange` callback.
+-   [Input] Make `onChange` callback as original `input.onChange`.
+
 ## [8.0.0-alpha.2] - 2019-04-19
 
 ### Components Added